+++ src/js/Form-model.js
@@ -655,9 +655,23 @@ define( function( require, exports, module ) {
      * @return {string}      new expression
      */
     FormModel.prototype.shiftRoot = function( expr ) {
+        var LITERALS = /"([^"]+)(")|'([^']+)(')/g;
         if ( this.hasInstance ) {
-            expr = expr.replace( /^(\/(?!model\/)[^\/][^\/\s]*\/)/g, '/model/instance[1]$1' );
-            expr = expr.replace( /([^a-zA-Z0-9\.\]\)\/\*_-])(\/(?!model\/)[^\/][^\/\s]*\/)/g, '$1/model/instance[1]$2' );
+            // Encode all string literals in order to exclude them, without creating a monsterly regex
+            expr = expr.replace( LITERALS, function( m, p1, p2, p3, p4 ) {
+                var encoded = p1 ? encodeURIComponent( p1 ) : encodeURIComponent( p3 );
+                var quote = p2 ? p2 : p4;
+                return quote + encoded + quote;
+            } );
+            // Insert /mode/instance[1]
+            expr = expr.replace( /^(\/(?!model\/)[^\/][^\/\s,"']*\/)/g, '/model/instance[1]$1' );
+            expr = expr.replace( /([^a-zA-Z0-9\.\]\)\/\*_-])(\/(?!model\/)[^\/][^\/\s,"']*\/)/g, '$1/model/instance[1]$2' );
+            // Decode string literals
+            expr = expr.replace( LITERALS, function( m, p1, p2, p3, p4 ) {
+                var decoded = p1 ? decodeURIComponent( p1 ) : decodeURIComponent( p3 );
+                var quote = p2 ? p2 : p4;
+                return quote + decoded + quote;
+            } );
         }
         return expr;
     };