+++ test/spec/formmodel.spec.js
@@ -448,9 +448,15 @@ describe( 'converting absolute paths', function() {
 
         // to leave unchanged
         [ 'path/to/node' ],
+        [ '"path/to/node"' ],
+        [ '\'path/to/node\'' ],
         [ 'concat(path/to/node, "2")' ],
         [ '../path/to/node' + '../node' ],
-        [ '/model/path/to/node' ]
+        [ '/model/path/to/node' ],
+        [ 'concat("/", "path/to/node")' ],
+        [ "concat('/path/to/node')" ],
+        [ 'concat("a", "[/path/to/node]")' ],
+        [ 'concat("\'", "/path/to/node", "\'")' ],
 
     ].forEach( function( test ) {
         it( 'converts correctly when the model and instance node are included in the model', function() {